add guava-31.1-jre.jar

+++ src/effectivejava/chapter8/item52/SetList.java
@@ -1,5 +1,9 @@
 package effectivejava.chapter8.item52;
-import java.util.*;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Set;
+import java.util.TreeSet;
 
 // What does this program print? (Page 241)
 public class SetList {
@@ -13,7 +17,7 @@ public class SetList {
         }
         for (int i = 0; i < 3; i++) {
             set.remove(i);
-            list.remove(i);
+            list.remove((Integer) i);
         }
         System.out.println(set + " " + list);
     }
